Rename go-redis
