+++ libs/go-redis/redis/redis_test.go
@@ -1,4 +1,4 @@
-package goredistrace
+package redis
 
 import (
 	"context"
